+++ indico/modules/rb/schemas.py
@@ -150,6 +150,14 @@ class ReservationSchema(mm.SQLAlchemyAutoSchema):
         return data
 
 
+class ReservationSchemaWithOwnership(ReservationSchema):
+    booked_for_self = Function(lambda r: session.user and r.booked_for_id == session.user.id)
+    booked_by_self = Function(lambda r: session.user and r.created_by_id == session.user.id)
+
+    class Meta(ReservationSchema.Meta):
+        fields = (*ReservationSchema.Meta.fields, 'booked_for_self', 'booked_by_self')
+
+
 class ReservationLinkedObjectDataSchema(mm.Schema):
     id = Number()
     title = Method('_get_title')
@@ -158,6 +166,8 @@ class ReservationLinkedObjectDataSchema(mm.Schema):
     event_url = Function(lambda obj: obj.event.url)
     own_room_id = Number()
     own_room_name = Function(lambda obj: (obj.own_room.name if obj.own_room else obj.own_room_name) or None)
+    start_dt = DateTime()
+    end_dt = DateTime()
 
     def _get_title(self, obj):
         if isinstance(obj, SessionBlock):
@@ -176,7 +186,8 @@ class ReservationUserEventSchema(mm.Schema):
 class ReservationOccurrenceLinkSchema(mm.SQLAlchemyAutoSchema):
     id = Number()
     type = EnumField(LinkType, attribute='link_type')
-    object = Nested(ReservationLinkedObjectDataSchema, only=('url', 'title', 'event_title', 'event_url'))
+    object = Nested(ReservationLinkedObjectDataSchema,
+                    only=('url', 'title', 'event_title', 'event_url', 'start_dt', 'end_dt'))
     start_dt = NaiveDateTime(attribute='reservation_occurrence.start_dt')
     state = EnumField(ReservationOccurrenceState, attribute='reservation_occurrence.state')
 
@@ -199,7 +210,11 @@ class ReservationOccurrenceSchema(mm.SQLAlchemyAutoSchema):
 
     class Meta:
         model = ReservationOccurrence
-        fields = ('start_dt', 'end_dt', 'is_valid', 'reservation', 'rejection_reason', 'state')
+        fields = ('start_dt', 'end_dt', 'is_valid', 'reservation', 'rejection_reason', 'state', 'link_id')
+
+
+class ReservationOccurrenceSchemaWithOwnership(ReservationOccurrenceSchema):
+    reservation = Nested(ReservationSchemaWithOwnership)
 
 
 class ReservationOccurrenceSchemaWithPermissions(ReservationOccurrenceSchema):
@@ -577,6 +592,7 @@ room_update_schema = RoomUpdateSchema()
 room_equipment_schema = RoomEquipmentSchema()
 map_areas_schema = MapAreaSchema(many=True)
 reservation_occurrences_schema = ReservationOccurrenceSchema(many=True)
+reservation_occurrences_schema_with_ownership = ReservationOccurrenceSchemaWithOwnership(many=True)
 reservation_occurrences_schema_with_permissions = ReservationOccurrenceSchemaWithPermissions(many=True)
 concurrent_pre_bookings_schema = ReservationConcurrentOccurrenceSchema(many=True)
 reservation_schema = ReservationSchema()